env-engr-ai/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyserial>=3.5
RPi.GPIO>=0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"
smbus2>=0.4.3
rich>=13.0.0
typer>=0.9.0
influxdb-client>=1.38.0
httpx>=0.25.0
